rename Assets to AssetsView, NeedsAndOffers to AssetsList, sharedComponents folder to components
