+++ client/src/features/static/policy/TrustAndSafety.tsx
@@ -5,7 +5,7 @@ import { Box, Grid, Container } from '@mui/material/';
 
 import type { Theme } from '@mui/material/styles';
 
-import SmallDisplayCard from '../../../assets/sharedComponents/SmallDisplayCard';
+import SmallDisplayCard from '../../../assets/components/SmallDisplayCard';
 
 const loremIpsum =
   'Lorem ipsum dolor sit amet, consectetur adipiscing elit. Facilisis placerat et, at vel tristique. Ac, gravida in quam gravida. Vel pretium nunc cursus donec enim. Sapien facilisis mauris justo, augue pharetra. Dignissim euismod fermentum sit gravida ut.';
